+++ src/Core/Dfe.Complete.Application/Projects/Queries/GetProject/GetProjectHistoryByProjectIdQuery.cs
@@ -3,6 +3,7 @@ using Dfe.Complete.Application.Common.Models;
 using Dfe.Complete.Application.Projects.Models;
 using AutoMapper;
 using Dfe.Complete.Application.Projects.Interfaces;
+using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Domain.ValueObjects;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -28,7 +29,7 @@ namespace Dfe.Complete.Application.Projects.Queries.GetProject
                     .Include(p => p.SignificantDateHistories)
                         .ThenInclude(ph => ph.User)
                     .Include(p => p.SignificantDateHistories)
-                        .ThenInclude(ph => ph.Reason)
+                        .ThenInclude(ph => ph.Reasons)
                     .FirstOrDefaultAsync(cancellationToken);
 
                 if (result!.SignificantDateHistories.Any(p => p.User == null))
@@ -39,6 +40,9 @@ namespace Dfe.Complete.Application.Projects.Queries.GetProject
                 result!.Notes = result.Notes.Where(n => n.NotableType == "SignificantDateHistoryReason").ToList();
 
                 var projectDto = mapper.Map<ProjectDto>(result);
+
+                // Added to prevent circular reference in integration test
+                projectDto.RegionalDeliveryOfficer = new User();
 
                 return Result<ProjectDto>.Success(projectDto);
             }